refactor: 수익률 계산 책임을 Lottos에서 LottoWinner로 이동 - Lottos에서 수익률을 계산할 경우, 당첨 통계를 인자로 넣어주어야 해서 비효율적 - 뿐만 아니라 LottoWinner로 수익률 계산 책임을 옮길 경우 Controller에서의 가독성도 향상(시간 흐름대로 코드 진행)
