+++ src/main/java/lotto/domain/LottoWinner.java
@@ -13,4 +13,17 @@ public class LottoWinner {
     public Map<LottoWinnerType, Integer> getWinners() {
         return Collections.unmodifiableMap(winners);
     }
+
+    public float calculateYield(int price) {
+        long profit = 0;
+        for (Map.Entry<LottoWinnerType, Integer> winner : winners.entrySet()) {
+            profit += calculateProfit(winner.getKey(), winner.getValue());
+        }
+
+        return (float) profit / price;
+    }
+
+    private int calculateProfit(LottoWinnerType winnerType, int countOfLotto) {
+        return LottoWinnerType.prize(winnerType) * countOfLotto;
+    }
 }
