+++ src/test/java/lotto/domain/LottoWinnerTest.java
@@ -1,0 +1,4 @@
+import static org.junit.jupiter.api.Assertions.*;
+class LottoWinnerTest {
+  
+}
